Bug fixes to spec Removed levmar from fitting New filter for CCD data based on stddev
git-svn-id: http://svn.code.sf.net/p/pyspec/code/trunk@229 d19dc245-ee4a-0410-ba2b-bd697e8c3f69

+++ pyspec/ccd/files.py
@@ -232,7 +232,7 @@ class PrincetonSPEFile():
         vxdim = self._readAtNumpy(14, 1, numpy.int16)[0]
         vydim = self._readAtNumpy(16, 1, numpy.int16)[0]
         dt = numpy.int16(self._readAtNumpy(108, 1, numpy.int16)[0])
-        data_types = (numpy.float, numpy.int32, numpy.int16, numpy.uint16)
+        data_types = (numpy.float32, numpy.int32, numpy.int16, numpy.uint16)
         if (dt > 3) or (dt < 0):
             raise Exception("Unknown data type")
         self._dataType = data_types[dt]
